refactor: generic data and dataset model
