+++ src/models/layer-model/common/data-model/utils/data-utils.ts
@@ -1,60 +1,31 @@
 import { getValue } from 'qlik-chart-modules';
 
 module DataUtils {
-  export function getAttributeMeta(dimensionInfo: any) {
-    let expressionMeta = {};
-    dimensionInfo.forEach((dimension: NxDimensionInfo, dimIndex: number) => {
-      dimension.qAttrExprInfo.forEach((attrExpr: NxAttrExprInfo, index: number) => {
-        expressionMeta[attrExpr.id] = {
-          id: attrExpr.id,
-          index,
-          dimIndex,
-          isDimension: false,
-          minValue: attrExpr.qMin,
-          maxValue: attrExpr.qMax,
-          title: attrExpr.qFallbackTitle,
-        };
-      });
+  export function getExpressionData(cell: NxCell, meta: any) {
+    let data = {};
+    for (const key in meta) {
+      const expressionMeta = meta[key];
+      const attribute = getAttribute(cell, expressionMeta);
+      data = { ...data, ...attribute };
+    }
 
-      dimension.qAttrDimInfo.forEach((dimExpr: NxAttrDimInfo, index: number) => {
-        expressionMeta[dimExpr.id] = {
-          id: dimExpr.id,
-          index,
-          dimIndex,
-          isDimension: true,
-          title: dimExpr.qFallbackTitle,
-        };
-      });
-    });
-
-    return expressionMeta;
+    return data;
   }
 
-  export function getAttributeData(cell: NxCell, meta: ExpressionMeta) {
+  export function getAttribute(cell: NxCell, meta: ExpressionMeta) {
     const attrValues = meta.isDimension ? cell.qAttrDims : cell.qAttrExps;
     if (!attrValues) {
       return {};
     }
     const qValue = getValue(attrValues, `qValues.${meta.index}`, {});
-    let value = null;
-    if (meta.isDimension) {
-      value = extractAttributeDimension(qValue);
-    } else {
-      value = extractAttributeExpression(qValue);
-    }
+    const value = extractValue(qValue, meta);
 
     return { [meta.id]: value };
   }
 
-  export function extractAttributeDimension(value: NxSimpleDimValue) {
-    if (value.qText) {
-      return value.qText as string;
-    }
-    return null;
-  }
-
-  export function extractAttributeExpression(value: NxSimpleValue) {
-    if (value.qNum !== undefined && value.qNum !== 'NaN') {
+  export function extractValue(value: NxSimpleValue, meta: ExpressionMeta) {
+    const hasMinOrMaxValue = !isNaN(meta.minValue as number) || !isNaN(meta.maxValue as number);
+    if (hasMinOrMaxValue && value.qNum !== undefined && value.qNum !== 'NaN') {
       return value.qNum as number;
     } else if (value.qText) {
       return value.qText as string;
