+++ src/models/layer-model/common/dataset-model/index.ts
@@ -1,11 +1,36 @@
-export abstract class DatasetModel {
-  abstract id: string;
-  abstract dataset: idevio.map.MemoryDataset | idevio.map.LocationDataset;
-  abstract loading: PromiseInterface;
+import GeneralUtils from '../../../../utils/general';
+import Utils from './utils';
+export class DatasetModel {
+  id: string;
+  layer: idevio.map.FeatureLayer;
+  geom: string;
+  dataset: idevio.map.MemoryDataset | idevio.map.LocationDataset;
+  loading: PromiseInterface;
   featureTable: any;
 
-  constructor() {
+  constructor(id: string, layer: idevio.map.FeatureLayer, geom: string = 'i:///pointgeom/default') {
+    this.id = id;
+    this.layer = layer;
+    this.geom = geom;
+    this.dataset = this.layer.getDataset();
+    this.loading = GeneralUtils.createPromise().resolve();
     this.featureTable = {};
+  }
+
+  update(crs: string, data: any) {
+    this.remove();
+    const info = Utils.getDatasetInfo(data);
+
+    if (info !== null) {
+      if (info.isGeoname) {
+        this.createLocationDataset(info.columns);
+        this.addLocationData(data, info.columns);
+      } else {
+        this.createMemoryDataset(crs);
+        this.addFeatureData(data);
+      }
+      this.layer.setDataset(this.dataset);
+    }
   }
 
   remove() {
@@ -14,20 +39,14 @@ export abstract class DatasetModel {
     }
   }
 
-  /**
-   * Can only be used after this.dataset is loaded.
-   */
-  hasFeatures() {
-    const features = this.dataset ? this.dataset.getAll() : null;
-    return features !== null && features?.length > 0;
+  createLocationDataset(columns: string[]) {
+    this.loading = GeneralUtils.createPromise();
+    const options = this.getDatasetOptions(columns) as idevio.map.RemoteServiceLocationDatasetOptions;
+    this.dataset = idevio.map.LocationDataset.create(this.geom, 'SERVICE', options);
   }
 
-  getBounds(): Rectangle | null {
-    if (this.hasFeatures()) {
-      return this.dataset.getAlternateBounds();
-    } else {
-      return null;
-    }
+  createMemoryDataset(crs: string) {
+    this.dataset = new idevio.map.MemoryDataset({ name: this.id, crs });
   }
 
   addLocationData(data: PointData[], dataOrder: string[]) {
@@ -42,7 +61,6 @@ export abstract class DatasetModel {
         dataOrder.forEach((key) => {
           featureData.push(row[key]);
         });
-        // geoname needs to be first
         collectedData.push(featureData);
       } catch {
         console.log('Failed lookup; ', row.geoname);
@@ -72,30 +90,6 @@ export abstract class DatasetModel {
     }
   }
 
-  getDatasetInfo(data: Data[]) {
-    for (const index in data) {
-      const row = data[index];
-      const columns = this.sortColumns(Object.keys(row));
-      if (row.hasOwnProperty('coords')) {
-        return {
-          columns,
-          isGeoname: false,
-        };
-      }
-      if (row.hasOwnProperty('geoname')) {
-        return {
-          columns,
-          isGeoname: true,
-        };
-      }
-    }
-    return null;
-  }
-
-  sortColumns(columns: string[]) {
-    return [...columns.filter((key) => key === 'geoname'), ...columns.filter((key) => key !== 'geoname')];
-  }
-
   getDatasetOptions(columnNames: string[]) {
     return {
       columnNames,
@@ -117,5 +111,21 @@ export abstract class DatasetModel {
       },
       onerror: () => {},
     };
+  }
+
+  /**
+   * Can only be used after this.dataset is loaded.
+   */
+  hasFeatures() {
+    const features = this.dataset ? this.dataset.getAll() : null;
+    return features !== null && features?.length > 0;
+  }
+
+  getBounds(): Rectangle | null {
+    if (this.hasFeatures()) {
+      return this.dataset.getAlternateBounds();
+    } else {
+      return null;
+    }
   }
 }
