+++ src/models/layer-model/point-layer/index.ts
@@ -1,33 +1,36 @@
 import { getValue } from 'qlik-chart-modules';
-import { PointLayerDatasetModel } from './models/dataset-model';
-import { PointLayerDataModel } from './models/data-model';
+import { DatasetModel } from '../common/dataset-model';
+import { DataModel } from '../common/data-model';
 import { SymbolModel } from './models/symbol-model';
 import { LayerModel } from '../common/layer-model';
-import LayoutService from './services/layout-service';
+import { layoutService as createLayoutService } from 'qlik-chart-modules';
 
 export class PointLayerModel extends LayerModel implements PointLayerModelInterface {
   id: string;
   layer: idevio.map.FeatureLayer;
-  dataModel: PointLayerDataModel;
-  datasetModel: PointLayerDatasetModel;
+  dataModel: DataModel;
+  datasetModel: DatasetModel;
   symbolModel: SymbolModel;
 
   constructor(mapModel: MapModelInterFace, id: string) {
     super(mapModel);
     this.id = id;
     this.layer = new idevio.map.FeatureLayer(mapModel.map, { drawOrder: 1001, pickable: true });
-    this.dataModel = new PointLayerDataModel();
-    this.datasetModel = new PointLayerDatasetModel(id, this.layer);
+    this.dataModel = new DataModel();
+    this.datasetModel = new DatasetModel(id, this.layer);
     this.symbolModel = new SymbolModel();
   }
 
   update(layout: PointLayerLayout) {
-    const layoutService = LayoutService.create(layout);
+    const layoutService = createLayoutService({
+      source: layout,
+    });
 
     this.dataModel.update(layoutService);
     const data = this.dataModel.getData();
+    const meta = this.dataModel.getMeta();
 
-    const collectedData = this.symbolModel.addSymbol(data, this.dataModel.meta, layoutService);
+    const collectedData = this.symbolModel.addSymbol(data, meta, layoutService);
     const crs = getValue(this.mapModel, 'baseMapModel.crs');
     this.datasetModel.update(crs, collectedData);
